Refactor out utils for resolving indices
